new server-specific daily tournament feature, new image art for shop and items, fixed bugs in coin flip function & account creation

+++ Utilities.py
@@ -5,7 +5,7 @@ class Utilities:
     def __init__(self, client):
         self.client = client
 
-    @commands.cooldown(1,  15, commands.BucketType.user)
+    @commands.cooldown(1,  25, commands.BucketType.user)
     @commands.command(name='clear', description='Deletes messages in the channel.',
                       brief='can use "=clear" or "=clear X", with X being #  messages to delete',
                       aliases=['c', 'clr', 'CLEAR', 'C', 'CLR', 'clean', 'CLEAN'], pass_context=True)
@@ -13,6 +13,9 @@ class Utilities:
         # try-catch block, because of *args array.
         # if no argument given in discord after "=clear", it will go to the exception
         try:
+            if int(args[0]) > 100:
+                await self.client.say("100 messages maximum!")
+                return
             deleted = await self.client.purge_from(context.message.channel, limit=int(args[0]))
             await self.client.say("Deleted %s message(s)" % str(len(deleted)))
 
@@ -35,7 +38,7 @@ class Utilities:
 
     @commands.cooldown(1, 5, commands.BucketType.user)
     @commands.command(name='invite', description='Give link to code', brief='can use "=code',
-                      aliases=['link', 'botlink', 'invitelink'], pass_context=True)
+                      aliases=['link', 'botlink', 'invitelink', 'INVITE'], pass_context=True)
     async def invite_link(self, context):
         # embed link in <> so discord will disable "link preview embedding"
         await self.client.say('Invite link for this **open-source** bot: \n'
